初始化版本： 1、部分bug修复 2、增加生产配置 3、用户名去除只限手机号注册 4、商品和订单支持小数价格

+++ src/main/java/ltd/newbee/mall/api/mall/vo/NewBeeMallGoodsDetailVO.java
@@ -12,6 +12,7 @@ import io.swagger.annotations.ApiModelProperty;
 import lombok.Data;
 
 import java.io.Serializable;
+import java.math.BigDecimal;
 
 /**
  * 商品详情页VO
@@ -32,7 +33,7 @@ public class NewBeeMallGoodsDetailVO implements Serializable {
     private String goodsCoverImg;
 
     @ApiModelProperty("商品价格")
-    private Integer sellingPrice;
+    private BigDecimal sellingPrice;
 
     @ApiModelProperty("商品标签")
     private String tag;
